requirements/dev.txt
-r base.txt
django-debug-toolbar==1.4
# Required by django-debug-toolbar
sqlparse==0.1.18

coverage==3.7.1
factory_boy==2.7.0
flake8==2.5.1
# Required by flake8
pyflakes==0.8.1
pep8==1.5.7
mccabe==0.3.1
mock==1.3.0
pylint==1.5.4
# Required by pylint
logilab-common==1.1.0
astroid==1.4.4
coveralls==1.1

# disable Fabric when using Python 3.3
# Fabric==1.8.2
# Required by Fabric
# paramiko==1.13.0
# pycrypto==2.6.1
# ecdsa==0.11

jupyter==1.0.0
